+++ bloom/commands/git/release.py
@@ -218,7 +218,10 @@ def execute_track(track, track_dict, release_inc, pretend=True):
     # setup extra settings
     archive_dir_path = tempfile.mkdtemp()
     settings['archive_dir_path'] = archive_dir_path
-    archive_file = '{name}-{release_tag}.tar.gz'.format(**settings)
+    if settings['release_tag'] != ':{none}':
+        archive_file = '{name}-{release_tag}.tar.gz'.format(**settings)
+    else:
+        archive_file = '{name}.tar.gz'.format(**settings)
     settings['archive_path'] = os.path.join(archive_dir_path, archive_file)
     # execute actions
     info("", use_prefix=False)